requirements.txt
line-bot-sdk
requests
openai
boto3
transformers
